Novel MBA method using Prime numbers
A matrix is created from X (the features) and y (a boolean array).
Each unique element is mapped to a unique prime number. Then each
itemset is summed to create a product vector. To check if item_no=2
is in our itemset, we check if each element of product vector is
divisible by 2. If we want to check if item_no 2 and 3 are both in
all our itmesets, we check if each element of product vector is
divisible by 2x3=6. Thus, checking for sets larger than one can
be done much easier.

The hope behind this approach is that it might be a bit faster since
it realies on matrix operations.